refactor: DTO패키지 구조 및 DTO명 변경
